pyproject.toml
[project]
name = "mini-plataforma-analise-dados"
version = "0.1.0"
description = "API-only PT-BR data analysis platform (CSV/XLSX → summary + NL Q&A)"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.136",
    "uvicorn[standard]>=0.46",
    "pydantic>=2.13",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.26",
    "pyjwt>=2.12",
    "pwdlib[bcrypt]>=0.3",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "structlog>=24.0",
    "python-dotenv>=1.0",
    "anyio>=4.0",
    # Phase 2-5 deps pre-declared so Docker image is complete from Phase 1:
    "pandas>=3.0",
    "openpyxl>=3.1",
    "duckdb>=1.5",
    "sqlglot>=30.6",
    "charset-normalizer>=3.4",
    "altair>=6.1",
    "openai>=2.32",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3",
    "httpx>=0.28",
    "respx>=0.23",
    "ruff",
    "mypy",
]
ui = [
    "streamlit>=1.40",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# Long signatures + canned LLM payloads are easier to read in one line
# inside tests; B (bugbear) gets noisy on intentional patterns.
"tests/**" = ["B", "E501"]

[tool.mypy]
python_version = "3.12"
strict = false
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: performance tests that process large fixtures or run full pipelines",
]
# Skip slow tests by default; run with `pytest -m slow` when validating SLAs.
addopts = "-m 'not slow'"
